agent_v0/requirements.txt
fastapi
openai
uvicorn
pydantic
python-dotenv
langgraph
langchain-core
langchain-community
requests
langchain-openai>=0.3.7
langchain-anthropic>=0.3.9
tavily-python>=0.5.0
langchain-groq>=0.2.4
arxiv>=2.1.3
pymupdf>=1.25.3
xmltodict>=0.14.2
linkup-sdk>=0.2.3
duckduckgo-search>=3.0.0
exa-py>=1.8.8
beautifulsoup4==4.13.3
langchain-deepseek>=0.1.2
